+++ plugins/a11yhelp/dialogs/lang/ko.js
@@ -79,6 +79,11 @@ CKEDITOR.plugins.setLang( 'a11yhelp', 'ko', {
 			{
 			name: ' 접근성 도움말',
 			legend: '${a11yHelp} 누르시오'
+		},
+			{
+			name: ' Paste as plain text', // MISSING
+			legend: 'Press ${pastetext}', // MISSING
+			legendEdge: 'Press ${pastetext}, followed by ${paste}' // MISSING
 		}
 		]
 	}